+++ crates/burn-jit/src/kernel/conv/conv2d.rs
@@ -1,261 +1,175 @@
+use std::marker::PhantomData;
+
 use burn_cube::{
-    cpa,
-    dialect::{ComputeShader, Elem, Scope, Variable, Visibility},
-    Compilation, CompilationInfo, CompilationSettings, EagerHandle, Execution, InputInfo,
-    OutputInfo, WorkgroupLaunch,
+    branch::*,
+    dialect::{ComputeShader, Elem, Visibility},
+    *,
 };
 use burn_tensor::{
     ops::{conv::calculate_conv_output_size, ConvOptions},
     Shape,
 };
-use std::marker::PhantomData;
 
 use crate::{
-    element::JitElement,
-    kernel::{into_contiguous, GpuComputeShaderPhase},
+    kernel::into_contiguous,
     ops::{
         numeric::{empty_device, zeros_device},
         reshape,
     },
     tensor::JitTensor,
-    JitRuntime,
+    FloatElement, JitRuntime,
 };
 
+#[cube]
+fn kernel<F: Float>(
+    input: Array<F>,
+    weight: Array<F>,
+    bias: Array<F>,
+    mut output: Array<F>,
+    conv_stride_0: UInt,
+    conv_stride_1: UInt,
+    dilation_0: UInt,
+    dilation_1: UInt,
+    padding_0: UInt,
+    padding_1: UInt,
+    groups: UInt,
+    kernel_size_0_unroll: Comptime<Option<UInt>>,
+    kernel_size_1_unroll: Comptime<Option<UInt>>,
+) {
+    let in_channels = shape::<F>(weight, 1u32);
+
+    let kernel_size_0 = Comptime::value_or(kernel_size_0_unroll, || shape::<F>(weight, 2u32));
+    let unroll_0 = Comptime::is_some(kernel_size_0_unroll);
+    let kernel_size_1 = Comptime::value_or(kernel_size_1_unroll, || shape::<F>(weight, 3u32));
+    let unroll_1 = Comptime::is_some(kernel_size_1_unroll);
+
+    let b = AbsoluteIndex::get() / stride::<F>(output, 0u32) % shape::<F>(output, 0u32);
+    let oc = AbsoluteIndex::get() / stride::<F>(output, 1u32) % shape::<F>(output, 1u32);
+    let oh = AbsoluteIndex::get() / stride::<F>(output, 2u32) % shape::<F>(output, 2u32);
+    let ow = AbsoluteIndex::get() / stride::<F>(output, 3u32) % shape::<F>(output, 3u32);
+
+    let g = (shape::<F>(weight, 0u32) + oc) % groups;
+    let ic_start = in_channels * g;
+    let ic_end = ic_start + in_channels;
+    let mut sum = bias[oc];
+
+    let ih_base = oh * conv_stride_0;
+    let iw_base = ow * conv_stride_1;
+
+    let weight_stride_1 = stride::<F>(weight, 1u32);
+    let weight_stride_2 = stride::<F>(weight, 2u32);
+    let weight_stride_3 = stride::<F>(weight, 3u32);
+
+    let input_stride_1 = stride::<F>(input, 1u32);
+    let input_stride_2 = stride::<F>(input, 2u32);
+    let input_stride_3 = stride::<F>(input, 3u32);
+    let input_shape_2 = shape::<F>(input, 2u32);
+    let input_shape_3 = shape::<F>(input, 3u32);
+
+    let border_top = padding_0;
+    let border_left = padding_1;
+    let border_bottom = input_shape_2 + padding_0;
+    let border_right = input_shape_3 + padding_1;
+
+    let index_input_0 = b * stride::<F>(input, 0u32);
+    let index_weight_0 = oc * stride::<F>(weight, 0u32);
+
+    for ic in range(ic_start, ic_end, Comptime::new(false)) {
+        let index_input_1 = ic * input_stride_1;
+        let index_weight_1 = (ic - ic_start) * weight_stride_1;
+
+        for kh in range(0u32, kernel_size_0, unroll_0) {
+            for kw in range(0u32, kernel_size_1, unroll_1) {
+                let ih = kh * dilation_0 + ih_base;
+                let iw = kw * dilation_1 + iw_base;
+
+                let within_padding = ih >= border_top
+                    && ih < border_bottom
+                    && iw >= border_left
+                    && iw < border_right;
+
+                if within_padding {
+                    let ih_pad = ih - padding_0;
+                    let iw_pad = iw - padding_1;
+
+                    let index_input = index_input_0
+                        + index_input_1
+                        + ih_pad * input_stride_2
+                        + iw_pad * input_stride_3;
+
+                    let index_weight = index_weight_0
+                        + index_weight_1
+                        + kh * weight_stride_2
+                        + kw * weight_stride_3;
+
+                    sum += input[index_input] * weight[index_weight];
+                }
+            }
+        }
+    }
+
+    output[AbsoluteIndex::get()] = sum;
+}
+
 #[derive(new)]
-struct Conv2dEagerKernel<R: JitRuntime, E: JitElement> {
+struct Conv2dEagerKernel<R: JitRuntime, E: FloatElement> {
+    kernel_size_0: Option<u32>,
+    kernel_size_1: Option<u32>,
     _runtime: PhantomData<R>,
     _elem: PhantomData<E>,
 }
 
-struct Conv2dComputeShader<E: JitElement> {
-    input: Variable,
-    weight: Variable,
-    bias: Variable,
-    output: Variable,
+struct Conv2dComputeShader<E: FloatElement> {
+    input: ExpandElement,
+    weight: ExpandElement,
+    bias: ExpandElement,
+    output: ExpandElement,
+    conv_stride_0: ExpandElement,
+    conv_stride_1: ExpandElement,
+    dilation_0: ExpandElement,
+    dilation_1: ExpandElement,
+    padding_0: ExpandElement,
+    padding_1: ExpandElement,
+    groups: ExpandElement,
+    kernel_size_0: Option<u32>,
+    kernel_size_1: Option<u32>,
     _elem: PhantomData<E>,
 }
 
-impl<E: JitElement> Conv2dComputeShader<E> {
-    fn expand(self, scope: &mut Scope) {
-        let input = self.input;
-        let weight = self.weight;
-        let bias = self.bias;
-        let output = self.output;
-        let id = Variable::Id;
-
-        let input_stride_0 = scope.create_local(Elem::UInt);
-        let input_stride_1 = scope.create_local(Elem::UInt);
-        let input_stride_2 = scope.create_local(Elem::UInt);
-        let input_stride_3 = scope.create_local(Elem::UInt);
-        let input_shape_0 = scope.create_local(Elem::UInt);
-        let input_shape_1 = scope.create_local(Elem::UInt);
-        let input_shape_2 = scope.create_local(Elem::UInt);
-        let input_shape_3 = scope.create_local(Elem::UInt);
-        cpa!(scope, input_stride_0 = stride(input, 0u32));
-        cpa!(scope, input_stride_1 = stride(input, 1u32));
-        cpa!(scope, input_stride_2 = stride(input, 2u32));
-        cpa!(scope, input_stride_3 = stride(input, 3u32));
-        cpa!(scope, input_shape_0 = shape(input, 0u32));
-        cpa!(scope, input_shape_1 = shape(input, 1u32));
-        cpa!(scope, input_shape_2 = shape(input, 2u32));
-        cpa!(scope, input_shape_3 = shape(input, 3u32));
-
-        let output_stride_0 = scope.create_local(Elem::UInt);
-        let output_stride_1 = scope.create_local(Elem::UInt);
-        let output_stride_2 = scope.create_local(Elem::UInt);
-        let output_stride_3 = scope.create_local(Elem::UInt);
-        let output_shape_0 = scope.create_local(Elem::UInt);
-        let output_shape_1 = scope.create_local(Elem::UInt);
-        let output_shape_2 = scope.create_local(Elem::UInt);
-        let output_shape_3 = scope.create_local(Elem::UInt);
-        cpa!(scope, output_stride_0 = stride(output, 0u32));
-        cpa!(scope, output_stride_1 = stride(output, 1u32));
-        cpa!(scope, output_stride_2 = stride(output, 2u32));
-        cpa!(scope, output_stride_3 = stride(output, 3u32));
-        cpa!(scope, output_shape_0 = shape(output, 0u32));
-        cpa!(scope, output_shape_1 = shape(output, 1u32));
-        cpa!(scope, output_shape_2 = shape(output, 2u32));
-        cpa!(scope, output_shape_3 = shape(output, 3u32));
-
-        let weight_stride_0 = scope.create_local(Elem::UInt);
-        let weight_stride_1 = scope.create_local(Elem::UInt);
-        let weight_stride_2 = scope.create_local(Elem::UInt);
-        let weight_stride_3 = scope.create_local(Elem::UInt);
-        let weight_shape_0 = scope.create_local(Elem::UInt);
-        let in_channels = scope.create_local(Elem::UInt);
-        let kernel_size_0 = scope.create_local(Elem::UInt);
-        let kernel_size_1 = scope.create_local(Elem::UInt);
-        cpa!(scope, weight_stride_0 = stride(weight, 0u32));
-        cpa!(scope, weight_stride_1 = stride(weight, 1u32));
-        cpa!(scope, weight_stride_2 = stride(weight, 2u32));
-        cpa!(scope, weight_stride_3 = stride(weight, 3u32));
-        cpa!(scope, weight_shape_0 = shape(weight, 0u32));
-        cpa!(scope, in_channels = shape(weight, 1u32));
-        cpa!(scope, kernel_size_0 = shape(weight, 2u32));
-        cpa!(scope, kernel_size_1 = shape(weight, 3u32));
-
-        let conv_stride_0 = Variable::GlobalScalar(0, Elem::UInt);
-        let conv_stride_1 = Variable::GlobalScalar(1, Elem::UInt);
-        let dilation_0 = Variable::GlobalScalar(2, Elem::UInt);
-        let dilation_1 = Variable::GlobalScalar(3, Elem::UInt);
-        let padding_0 = Variable::GlobalScalar(4, Elem::UInt);
-        let padding_1 = Variable::GlobalScalar(5, Elem::UInt);
-        let groups = Variable::GlobalScalar(6, Elem::UInt);
-
-        let b = scope.create_local(Elem::UInt);
-        let oc = scope.create_local(Elem::UInt);
-        let oh = scope.create_local(Elem::UInt);
-        let ow = scope.create_local(Elem::UInt);
-        let g = scope.create_local(Elem::UInt);
-
-        let ic_start = scope.create_local(Elem::UInt);
-        let ic_end = scope.create_local(Elem::UInt);
-
-        cpa!(scope, b = id / output_stride_0);
-        cpa!(scope, b = b % output_shape_0);
-
-        cpa!(scope, oc = id / output_stride_1);
-        cpa!(scope, oc = oc % output_shape_1);
-
-        cpa!(scope, oh = id / output_stride_2);
-        cpa!(scope, oh = oh % output_shape_2);
-
-        cpa!(scope, ow = id / output_stride_3);
-        cpa!(scope, ow = ow % output_shape_3);
-
-        cpa!(scope, g = weight_shape_0 + oc);
-        cpa!(scope, g = g % groups);
-
-        cpa!(scope, ic_start = in_channels * g);
-        cpa!(scope, ic_end = ic_start + in_channels);
-
-        let sum = scope.create_local(output.item());
-        cpa!(scope, sum = bias[oc]);
-
-        let ih_base = scope.create_local(Elem::UInt);
-        let iw_base = scope.create_local(Elem::UInt);
-        let ih = scope.create_local(Elem::UInt);
-        let iw = scope.create_local(Elem::UInt);
-
-        let padding = scope.create_local(Elem::Bool);
-        let padding_accumulator = scope.create_local(Elem::Bool);
-        let border_top = scope.create_local(Elem::UInt);
-        let border_bottom = scope.create_local(Elem::UInt);
-        let border_left = scope.create_local(Elem::UInt);
-        let border_right = scope.create_local(Elem::UInt);
-
-        let ih_pad = scope.create_local(Elem::UInt);
-        let iw_pad = scope.create_local(Elem::UInt);
-
-        let index_input = scope.create_local(Elem::UInt);
-        let index_input_0 = scope.create_local(Elem::UInt);
-        let index_input_1 = scope.create_local(Elem::UInt);
-        let index_input_2 = scope.create_local(Elem::UInt);
-        let index_input_3 = scope.create_local(Elem::UInt);
-
-        let index_weight = scope.create_local(Elem::UInt);
-        let index_weight_0 = scope.create_local(Elem::UInt);
-        let index_weight_1 = scope.create_local(Elem::UInt);
-        let index_weight_2 = scope.create_local(Elem::UInt);
-        let index_weight_3 = scope.create_local(Elem::UInt);
-
-        let input_value = scope.create_local(input.item());
-        let weight_value = scope.create_local(weight.item());
-        let value_product = scope.create_local(input.item());
-
-        cpa!(scope, ih_base = oh * conv_stride_0);
-        cpa!(scope, iw_base = ow * conv_stride_1);
-
-        cpa!(scope, border_top = padding_0);
-        cpa!(scope, border_left = padding_1);
-        cpa!(scope, border_bottom = input_shape_2 + padding_0);
-        cpa!(scope, border_right = input_shape_3 + padding_1);
-
-        cpa!(scope, index_input_0 = b * input_stride_0);
-        cpa!(scope, index_weight_0 = oc * weight_stride_0);
-
-        cpa!(
-            scope,
-            range(ic_start, ic_end).for_each(|ic, scope| {
-                cpa!(scope, index_input_1 = ic * input_stride_1);
-                cpa!(scope, index_weight_1 = ic - ic_start);
-                cpa!(scope, index_weight_1 *= weight_stride_1);
-
-                cpa!(
-                    scope,
-                    range(0u32, kernel_size_0).for_each(|kh, scope| {
-                        cpa!(
-                            scope,
-                            range(0u32, kernel_size_1).for_each(|kw, scope| {
-                                cpa!(scope, ih = kh * dilation_0);
-                                cpa!(scope, ih += ih_base);
-                                cpa!(scope, iw = kw * dilation_1);
-                                cpa!(scope, iw += iw_base);
-
-                                cpa!(scope, padding_accumulator = ih >= border_top);
-                                cpa!(scope, padding = ih < border_bottom);
-                                cpa!(scope, padding_accumulator = padding_accumulator && padding);
-                                cpa!(scope, padding = iw >= border_left);
-                                cpa!(scope, padding_accumulator = padding_accumulator && padding);
-                                cpa!(scope, padding = iw < border_right);
-                                cpa!(scope, padding_accumulator = padding_accumulator && padding);
-
-                                cpa!(scope, if(padding_accumulator).then(|scope|{
-                                    cpa!(scope, ih_pad = ih - padding_0);
-                                    cpa!(scope, iw_pad = iw - padding_1);
-
-                                    cpa!(scope, index_input_2 = ih_pad * input_stride_2);
-                                    cpa!(scope, index_input_3 = iw_pad * input_stride_3);
-
-                                    cpa!(scope, index_input = index_input_0);
-                                    cpa!(scope, index_input += index_input_1);
-                                    cpa!(scope, index_input += index_input_2);
-                                    cpa!(scope, index_input += index_input_3);
-
-                                    cpa!(scope, index_weight_2 = kh * weight_stride_2);
-                                    cpa!(scope, index_weight_3 = kw * weight_stride_3);
-
-                                    cpa!(scope, index_weight = index_weight_0);
-                                    cpa!(scope, index_weight += index_weight_1);
-                                    cpa!(scope, index_weight += index_weight_2);
-                                    cpa!(scope, index_weight += index_weight_3);
-
-                                    cpa!(scope, input_value = input[index_input]);
-                                    cpa!(scope, weight_value = weight[index_weight]);
-                                    cpa!(scope, value_product = input_value * weight_value);
-                                    cpa!(scope, sum += value_product);
-                                }));
-                            })
-                        );
-                    })
-                );
-            })
-        );
-
-        cpa!(scope, output[id] = sum);
-    }
-}
-
-impl<R: JitRuntime, E: JitElement> GpuComputeShaderPhase for Conv2dEagerKernel<R, E> {
+impl<R: JitRuntime, E: FloatElement> GpuComputeShaderPhase for Conv2dEagerKernel<R, E> {
     fn compile(&self) -> ComputeShader {
-        let mut scope = Scope::root();
+        let mut context = CubeContext::root();
         let item = E::cube_elem().into();
 
-        let input = Variable::GlobalInputArray(0, item);
-        let weight = Variable::GlobalInputArray(1, item);
-        let bias = Variable::GlobalInputArray(2, item);
-        let output = Variable::GlobalOutputArray(0, item);
-
-        scope.write_global_custom(output);
+        let input = context.input(0, item);
+        let weight = context.input(1, item);
+        let bias = context.input(2, item);
+        let output = context.output(0, item);
+        let conv_stride_0 = context.scalar(0, Elem::UInt);
+        let conv_stride_1 = context.scalar(1, Elem::UInt);
+        let dilation_0 = context.scalar(2, Elem::UInt);
+        let dilation_1 = context.scalar(3, Elem::UInt);
+        let padding_0 = context.scalar(4, Elem::UInt);
+        let padding_1 = context.scalar(5, Elem::UInt);
+        let groups = context.scalar(6, Elem::UInt);
 
         Conv2dComputeShader {
             input,
             weight,
             bias,
             output,
+            conv_stride_0,
+            conv_stride_1,
+            dilation_0,
+            dilation_1,
+            padding_0,
+            padding_1,
+            groups,
+            kernel_size_0: self.kernel_size_0,
+            kernel_size_1: self.kernel_size_1,
             _elem: PhantomData::<E>,
         }
-        .expand(&mut scope);
+        .expand(&mut context);
 
         let input = InputInfo::Array {
             item,
@@ -276,6 +190,7 @@ impl<R: JitRuntime, E: JitElement> GpuComputeShaderPhase for Conv2dEagerKernel<R
 
         let output = OutputInfo::Array { item };
 
+        let scope = context.into_scope();
         let info = CompilationInfo {
             inputs: vec![input, weight, bias, scalars],
             outputs: vec![output],
@@ -287,11 +202,37 @@ impl<R: JitRuntime, E: JitElement> GpuComputeShaderPhase for Conv2dEagerKernel<R
     }
 
     fn id(&self) -> String {
-        format!("{:?}", core::any::TypeId::of::<Self>(),)
+        format!(
+            "{:?}-{:?}-{:?}",
+            core::any::TypeId::of::<Self>(),
+            self.kernel_size_0,
+            self.kernel_size_1
+        )
     }
 }
 
-pub(crate) fn conv2d<R: JitRuntime, E: JitElement>(
+impl<E: FloatElement> Conv2dComputeShader<E> {
+    fn expand(self, context: &mut CubeContext) {
+        kernel_expand::<E::CubeElement>(
+            context,
+            self.input,
+            self.weight,
+            self.bias,
+            self.output,
+            self.conv_stride_0,
+            self.conv_stride_1,
+            self.dilation_0,
+            self.dilation_1,
+            self.padding_0,
+            self.padding_1,
+            self.groups,
+            self.kernel_size_0.map(UInt::new),
+            self.kernel_size_1.map(UInt::new),
+        )
+    }
+}
+
+pub(crate) fn conv2d<R: JitRuntime, E: FloatElement>(
     input: JitTensor<R, E, 4>,
     weight: JitTensor<R, E, 4>,
     bias: Option<JitTensor<R, E, 1>>,
@@ -336,7 +277,8 @@ pub(crate) fn conv2d<R: JitRuntime, E: JitElement>(
         }
     };
 
-    let kernel = Conv2dEagerKernel::<R, E>::new();
+    let kernel = Conv2dEagerKernel::<R, E>::new(Some(kernel_0 as u32), Some(kernel_1 as u32));
+    // let kernel = Conv2dEagerKernel::<R, E>::new(None, None);
 
     Execution::start(kernel, input.client)
         .inputs(&[
