+++ crates/burn-jit/src/kernel/conv/mod.rs
@@ -1,7 +1,5 @@
 mod conv2d;
-mod conv_cube;
 mod conv_transpose2d;
 
 pub(crate) use conv2d::*;
-// pub(crate) use conv_cube::*;
 pub(crate) use conv_transpose2d::*;
